requirements.txt
git+https://github.com/petrpechman/m2scorer.git
pandas
ufal.udpipe
